Correct DTO data types

+++ backend/src/main/java/com/propertystack/homematch/dto/ListingDTO.java
@@ -12,9 +12,9 @@ public class ListingDTO {
     Long id;
     String address;
     BigDecimal price;
-    BigDecimal sqft;
+    Integer sqft;
     Integer beds;
-    BigDecimal baths;
+    Double baths;
 
     String listingUrl;
